Add a standard Blazor event callback to the component
